requirements.txt
Babel==2.6.0
Django==2.0.7
django-appconf==1.0.2
django-bootstrap4==0.0.6
django-imagekit==4.0.2
django-js-asset==1.1.0
django-mptt==0.9.0
django-phonenumber-field==2.0.0
phonenumberslite==8.9.10
pilkit==2.0
Pillow==5.2.0
pytz==2018.5
six==1.11.0
